Handle some Various Artists songs better
Some songs on Various Artists albums have no artistId field. In
this case (for SubSonic - iirc MadSonic does it a bit differently..)
the artist field is populated with the string from the metadata.

=>
  * Take Artist field from metadata when the artistId is missing

This should sync more Various Artists songs. However an Album
with 'Various Artists' as the Artist will still not be synced.

Why? If you fetch the album list for 'Various Artist' it is empty.
But please double check this, I did a quick N=1 check.